# docs/openapi.yml
openapi: 3.0.4
info:
  title: Concert Reservation API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://v1.api.concert.com
    description: Production server
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message
    ConcertDate:
      type: object
      properties:
        date:
          type: string
          format: date
        remainingSeats:
          type: integer
      required:
        - date
        - remainingSeats
    ConcertSeat:
      type: object
      properties:
        section:
          type: string
        row:
          type: string
        number:
          type: integer
        status:
          type: string
          enum: [AVAILABLE, TEMP_HOLD, SOLD]
      required:
        - section
        - row
        - number
        - status
...
paths:
...
  /concerts/{id}/dates:
    get:
      summary: Get available dates for a concert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Available dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConcertDate'
              example:
                - date: "2025-12-31"
                  remainingSeats: 150
        '404':
          description: Concert not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /concerts/{id}/dates/{date}/seats:
    get:
      summary: Get seats for a concert on a specific date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of seats with status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConcertSeat'
              example:
                - section: "VIP"
                  row: "A"
                  number: 1
                  status: AVAILABLE
                - section: "VIP"
                  row: "A"
                  number: 2
                  status: TEMP_HOLD
                - section: "R"
                  row: "B"
                  number: 10
                  status: SOLD
        '404':
          description: Concert or date not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...